+++ google-cloud-bigtable-stats/pom.xml
@@ -51,15 +51,11 @@
         <dependency>
             <groupId>com.google.cloud</groupId>
             <artifactId>google-cloud-monitoring</artifactId>
-            <version>3.2.9</version>
+            <version>3.3.1</version>
         </dependency>
         <dependency>
             <groupId>io.opencensus</groupId>
             <artifactId>opencensus-api</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>io.opencensus</groupId>
-            <artifactId>opencensus-impl</artifactId>
         </dependency>
         <dependency>
             <groupId>io.opencensus</groupId>
